Backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
numpy==2.2.2
opencv-python-headless==4.10.0.84
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
django-cors-headers==4.6.0
dj-database-url==2.3.0